Assign custom instrumentation method to `globalThis.fetch` in `setupSentry `script
