+++ shared/lib/trace.ts
@@ -378,61 +378,6 @@ function sentryStartSpanManual<T>(
   return actual(spanOptions, callback);
 }
 
-export const fetchWithSentryInstrumentation = async (
-  inputUrl: string | RequestInfo | URL,
-  opts?: RequestInit,
-) => {
-  const url =
-    typeof inputUrl === 'string'
-      ? inputUrl
-      : inputUrl.toString() || String(inputUrl);
-  const { method = 'GET' } = opts ?? {};
-
-  const response = await globalThis.fetch(url, {
-    method,
-    ...opts,
-  });
-
-  // Do not create spans for outgoing requests to a 'sentry.io' domain.
-  if (url.match(/^https?:\/\/([\w\d.@-]+\.)?sentry\.io(\/|$)/u)) {
-    return response;
-  }
-
-  try {
-    Sentry.startSpan(
-      { op: 'http.client', name: `${method} ${url}` },
-      (span) => {
-        const parsedURL = new URL(url, location.origin);
-
-        span.setAttributes({
-          'http.request.method': method,
-          'server.address': parsedURL.hostname,
-          'server.port': parsedURL.port || undefined,
-          'http.response.status_code': response.status,
-          'http.response_content_length': Number(
-            response.headers?.get('content-length'),
-          ),
-        });
-
-        const cloudflareRayId =
-          response.headers?.get('CF-RAY') ??
-          response.headers?.get('CF-Ray') ??
-          response.headers?.get('CF-ray') ??
-          response.headers?.get('cf_ray');
-        if (cloudflareRayId !== null) {
-          span.setAttribute('CF-Ray', cloudflareRayId);
-          const scope = Sentry.getCurrentScope();
-          scope?.setTag('CF-Ray', cloudflareRayId);
-        }
-      },
-    );
-  } catch (error) {
-    log(error);
-  }
-
-  return response;
-};
-
 function sentryWithIsolationScope<T>(callback: (scope: Sentry.Scope) => T): T {
   const actual = globalThis.sentry?.withIsolationScope;
