+++ shared/modules/fetch-with-timeout.ts
@@ -1,6 +1,5 @@
 import { memoize } from 'lodash';
 import { SECOND } from '../constants/time';
-import { fetchWithSentryInstrumentation } from '../lib/trace';
 
 /**
  * Returns a function that can be used to make an HTTP request but timing out
@@ -33,7 +32,7 @@ const getFetchWithTimeout = memoize((timeout = SECOND * 30) => {
     const abortHandler = () => combinedAbortController.abort();
     abortSignals.forEach((sig) => sig.addEventListener('abort', abortHandler));
 
-    const f = fetchWithSentryInstrumentation(url, {
+    const f = globalThis.fetch(url, {
       ...opts,
       signal: combinedAbortController.signal,
     });
